Media queries jay1

+++ react-app/src/components/Bookings/BookedTasks.js
@@ -4,6 +4,7 @@ import {editTaskThunk, getTasksThunk} from '../../store/tasks'
 import {editBookingThunk, deleteBookingThunk} from '../../store/booking'
 import '../Profile/ProfileTaskCard.css';
 import ProfileTaskCard from '../Profile/ProfileTaskCard';
+import TaskCard from '../Tasks/taskCard/taskCard';
 
 function BookedTasks({ taskId, booking, reviewArr }) {
   const task = useSelector(state => state.tasks[taskId])
@@ -46,29 +47,29 @@ function BookedTasks({ taskId, booking, reviewArr }) {
 
   const deleteHandler = async(e) => {
     try {
-        e.preventDefault();
+        if (window.confirm('Are you sure you want to drop this task?')) {
+          e.preventDefault();
 
-        const payload = {
-        ...task,
-        available: true,
-        tags
+          const payload = {
+          ...task,
+          available: true,
+          tags
+          }
+          await dispatch(editTaskThunk(payload))
+          await dispatch(deleteBookingThunk(booking))
         }
-        const editedTask = await dispatch(editTaskThunk(payload))
-        await dispatch(deleteBookingThunk(booking))
-
-        if(editedTask) window.alert("nice");
     }catch(err) {
         setValidationErrors(err.errors)
     }
   }
 
   return (
-    <div>
+    <div className='prof-taskcard-div'>
       {validationErrors && validationErrors.length > 0 && validationErrors.map(error => {
         return <div>{error}</div>
       })}
       {task && booking && (
-      <div>
+      <div className='prof-taskcard-div'>
         <ProfileTaskCard task={task} booking={booking} submitHandler={submitHandler} deleteHandler={deleteHandler} leftReview={leftReview} taskId={taskId}/>
       </div>
       )}
