+++ react-app/src/components/Reviews/SingleReview.js
@@ -1,17 +1,18 @@
 import React, { useState } from 'react';
+import { Link } from 'react-router-dom';
 import EditReviewForm from './EditReviewForm';
 import { useDispatch, useSelector } from 'react-redux';
 import { deleteReviewThunk } from '../../store/review'
 import EditReviewFormModal from './EditFormModal';
-//import "./SingleReview.css"
+import "./SingleReview.css"
 
 function SingleReview({ review }) {
   const dispatch = useDispatch();
-  const sessionUser = useSelector((state) => state.session.user)
+  const sessionUser = useSelector((state) => state.session.user);
   const [showEditForm, setShowEditForm] = useState(false);
 
   const deleteHandler = async () => {
-    await dispatch(deleteReviewThunk(review))
+    if (window.confirm('Are you sure you want to delete this review?')) await dispatch(deleteReviewThunk(review));
   }
 
   return (
@@ -22,8 +23,9 @@ function SingleReview({ review }) {
             <div className='profile-review'>
               <div className='review-rating'>Rating: {review.rating}</div>
               <div className='review-comment'>Comment: {review.comment}</div>
-              <div>
-                <EditReviewFormModal taskId={sessionUser.id} review={review} />
+              <Link className='view-task' to={`/tasks/${review.task_id}`}>View Task</Link>
+              <div className='review-profile-btn-div'>
+                <EditReviewFormModal review={review} />
                 <button className='review-profile-btns' onClick={deleteHandler}>Delete</button>
               </div>
             </div>
